+++ mosaic-runtime/src/commonMain/kotlin/com/jakewharton/mosaic/ansi.kt
@@ -11,20 +11,29 @@ internal const val CSI = "$ESC["
 internal const val OSC = "$ESC]"
 internal const val APC = "${ESC}_"
 
-internal const val modeFocus = 1004
-internal const val modeSynchronizedRendering = 2026
-internal const val modeSystemTheme = 2031
-internal const val modeInBandResize = 2048
+internal const val cursorMode = 25
+internal const val cursorEnable = "$CSI?${cursorMode}h"
+internal const val cursorDisable = "$CSI?${cursorMode}l"
 
-internal const val ansiBeginSynchronizedUpdate = "$CSI?${modeSynchronizedRendering}h"
-internal const val ansiEndSynchronizedUpdate = "$CSI?${modeSynchronizedRendering}l"
+internal const val focusMode = 1004
+internal const val focusEnable = "$CSI?${focusMode}h"
+internal const val focusDisable = "$CSI?${focusMode}l"
+
+internal const val synchronizedRenderingMode = 2026
+internal const val synchronizedRenderingEnable = "$CSI?${synchronizedRenderingMode}h"
+internal const val synchronizedRenderingDisable = "$CSI?${synchronizedRenderingMode}l"
+
+internal const val systemThemeMode = 2031
+internal const val systemThemeEnable = "$CSI?${systemThemeMode}h"
+internal const val systemThemeDisable = "$CSI?${systemThemeMode}l"
+
+internal const val inBandResizeMode = 2048
+internal const val inBandResizeEnable = "$CSI?${inBandResizeMode}h"
+internal const val inBandResizeDisable = "$CSI?${inBandResizeMode}l"
 
 internal const val ansiReset = "${CSI}0"
 internal const val clearLine = "${CSI}K"
 internal const val cursorUp = "${CSI}F"
-
-internal const val cursorHide = "$CSI?25l"
-internal const val cursorShow = "$CSI?25h"
 
 internal const val ansiSeparator = ";"
 internal const val ansiClosingCharacter = "m"