+++ mosaic-runtime/src/commonTest/kotlin/com/jakewharton/mosaic/CounterTest.kt
@@ -32,13 +32,13 @@ class CounterTest {
 
 	@Test fun counterInTerminalCenter() = runTest {
 		runMosaicTest {
-			terminalState.value = Terminal(size = IntSize(width = 30, height = 1))
+			terminalState.update { copy(size = IntSize(width = 30, height = 1)) }
 			setCounterInTerminalCenter()
 			for (count in 0..9) {
 				assertThat(awaitSnapshot()).isEqualTo("        The count is: $count       ")
 			}
 
-			terminalState.value = Terminal(size = IntSize(width = 20, height = 1))
+			terminalState.update { copy(size = IntSize(width = 20, height = 1)) }
 
 			// After changing the terminal size, we wait for the counter to increase before getting a
 			// new snapshot, otherwise there will be the previous value (9) and a different output size.